add to Inspect documentation
