+++ tests/util/sandbox/test_sandbox_setup.py
@@ -8,8 +8,7 @@ from inspect_ai.dataset import Sample
 from inspect_ai.model import ModelOutput
 from inspect_ai.scorer import CORRECT, includes
 from inspect_ai.solver import Generate, Solver, TaskState, solver
-from inspect_ai.util import sandbox
-from inspect_ai.util._sandbox.docker.config import is_dockerfile
+from inspect_ai.util import is_dockerfile, sandbox
 
 SANDBOX_SETUP_FILE = (Path(__file__).parent / "sandbox_setup.sh").as_posix()
 SANDBOX_SETUP_ERROR_FILE = (Path(__file__).parent / "sandbox_setup_error.sh").as_posix()
